Refactor view methods in service
