+++ safe_relay_service/relay/models.py
@@ -1,19 +1,12 @@
-from typing import Dict, Iterable, List, Union
-
 from django.contrib.postgres.fields import ArrayField
 from django.db import models
 
 from model_utils.models import TimeStampedModel
 
 from gnosis.eth import EthereumServiceProvider
-from gnosis.eth.constants import NULL_ADDRESS
 from gnosis.eth.django.models import (EthereumAddressField, Sha3HashField,
                                       Uint256Field)
-from gnosis.safe.safe_service import SafeOperation, SafeServiceException
-
-from safe_relay_service.gas_station.gas_station import GasStationProvider
-
-from .relay_service import RelayServiceException, RelayServiceProvider
+from gnosis.safe.safe_service import SafeOperation, SafeServiceProvider
 
 
 class SafeContract(TimeStampedModel):
@@ -21,10 +14,10 @@ class SafeContract(TimeStampedModel):
     master_copy = EthereumAddressField()
 
     def has_valid_code(self) -> bool:
-        return RelayServiceProvider().check_proxy_code(self.address)
+        return SafeServiceProvider().check_proxy_code(self.address)
 
     def has_valid_master_copy(self) -> bool:
-        return RelayServiceProvider().check_master_copy(self.address)
+        return SafeServiceProvider().check_master_copy(self.address)
 
     def get_balance(self, block_identifier=None):
         return EthereumServiceProvider().get_balance(address=self.address, block_identifier=block_identifier)
@@ -33,53 +26,7 @@ class SafeContract(TimeStampedModel):
         return self.address
 
 
-class SafeCreationManager(models.Manager):
-    def create_safe_tx(self, s: int, owners: Iterable[str], threshold: int, payment_token: Union[str, None],
-                       payment_token_eth_value: float=1.0, fixed_creation_cost: Union[int, None]=None):
-        """
-        Create models for safe tx
-        :param s: Random s value for ecdsa signature
-        :param owners: Owners of the new Safe
-        :param threshold: Minimum number of users required to operate the Safe
-        :param payment_token: Address of the payment token, if ether is not used
-        :param payment_token_eth_value: Value of payment_token per 1 ether
-        :param fixed_creation_cost: Fixed creation cost of Safe (Wei)
-        :rtype: SafeCreation
-        """
-
-        relay_service = RelayServiceProvider()
-        gas_station = GasStationProvider()
-        fast_gas_price: int = gas_station.get_gas_prices().fast
-        safe_creation_tx = relay_service.build_safe_creation_tx(s, owners, threshold, fast_gas_price, payment_token,
-                                                                payment_token_eth_value=payment_token_eth_value,
-                                                                fixed_creation_cost=fixed_creation_cost)
-
-        safe_contract = SafeContract.objects.create(address=safe_creation_tx.safe_address,
-                                                    master_copy=safe_creation_tx.master_copy)
-
-        return super().create(
-            deployer=safe_creation_tx.deployer_address,
-            safe=safe_contract,
-            master_copy=safe_creation_tx.master_copy,
-            funder=safe_creation_tx.funder,
-            owners=owners,
-            threshold=threshold,
-            payment=safe_creation_tx.payment,
-            tx_hash=safe_creation_tx.tx_hash.hex(),
-            gas=safe_creation_tx.gas,
-            gas_price=safe_creation_tx.gas_price,
-            payment_token=None if safe_creation_tx.payment_token == NULL_ADDRESS else safe_creation_tx.payment_token,
-            value=safe_creation_tx.tx_pyethereum.value,
-            v=safe_creation_tx.v,
-            r=safe_creation_tx.r,
-            s=safe_creation_tx.s,
-            data=safe_creation_tx.tx_pyethereum.data,
-            signed_tx=safe_creation_tx.tx_raw
-        )
-
-
 class SafeCreation(TimeStampedModel):
-    objects = SafeCreationManager()
     deployer = EthereumAddressField(primary_key=True)
     safe = models.OneToOneField(SafeContract, on_delete=models.CASCADE)
     master_copy = EthereumAddressField()
@@ -171,83 +118,9 @@ class SafeFunding(TimeStampedModel):
 
 
 class SafeMultisigTxManager(models.Manager):
-    class SafeMultisigTxExists(Exception):
-        pass
-
-    class SafeMultisigTxError(Exception):
-        pass
-
     def get_last_nonce_for_safe(self, safe_address: str):
         tx = self.filter(safe=safe_address).order_by('-nonce').first()
         return tx.nonce if tx else None
-
-    def create_multisig_tx(self,
-                           safe_address: str,
-                           to: str,
-                           value: int,
-                           data: bytes,
-                           operation: int,
-                           safe_tx_gas: int,
-                           data_gas: int,
-                           gas_price: int,
-                           gas_token: str,
-                           refund_receiver: str,
-                           nonce: int,
-                           signatures: List[Dict[str, int]]):
-        """
-        :return: Database model of SafeMultisigTx
-        :raises: SafeMultisigTxExists: If Safe Multisig Tx with nonce already exists
-        :raises: SafeMultisigTxError: If Safe Tx is not valid (not sorted owners, bad signature, bad nonce...)
-        """
-
-        if self.filter(safe=safe_address, nonce=nonce).exists():
-            raise self.SafeMultisigTxExists
-
-        relay_service = RelayServiceProvider()
-
-        signature_pairs = [(s['v'], s['r'], s['s']) for s in signatures]
-        signatures_packed = relay_service.signatures_to_bytes(signature_pairs)
-        safe_tx_hash = relay_service.get_hash_for_safe_tx(safe_address, to, value, data,
-                                                          operation, safe_tx_gas, data_gas, gas_price,
-                                                          gas_token, refund_receiver, nonce)
-
-        try:
-            tx_hash, tx = relay_service.send_multisig_tx(
-                safe_address,
-                to,
-                value,
-                data,
-                operation,
-                safe_tx_gas,
-                data_gas,
-                gas_price,
-                gas_token,
-                refund_receiver,
-                signatures_packed
-            )
-        except (SafeServiceException, RelayServiceException) as exc:
-            raise self.SafeMultisigTxError(str(exc)) from exc
-
-        safe_contract = SafeContract.objects.get(address=safe_address)
-
-        return super().create(
-            safe=safe_contract,
-            to=to,
-            value=value,
-            data=data,
-            operation=operation,
-            safe_tx_gas=safe_tx_gas,
-            data_gas=data_gas,
-            gas_price=gas_price,
-            gas_token=None if gas_token == NULL_ADDRESS else gas_token,
-            refund_receiver=refund_receiver,
-            nonce=nonce,
-            signatures=signatures_packed,
-            gas=tx['gas'],
-            safe_tx_hash=safe_tx_hash,
-            tx_hash=tx_hash.hex(),
-            tx_mined=False
-        )
 
 
 class SafeMultisigTx(TimeStampedModel):
